requirements.txt
Django==1.8.6
djangorestframework
python-dateutil
markdown2
beautifulsoup4
Pygments
django-allauth==0.29.0


# Django test supports
# ====================
tox
django_dynamic_fixture
django_nose
pytest
pytest-django
pytest-xdist
pytest-cov
nose-progressive
responses>=0.4.0
git+git://github.com/spulec/freezegun


# Django docs supports
# ====================
sphinx
recommonmark
sphinx-git
sphinx_rtd_theme

# django-accounts specific packages
# =============================
# Add here ...
